fixed comments

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoHibernateImpl.java
@@ -2,10 +2,11 @@ package jm.task.core.jdbc.dao;
 
 import jm.task.core.jdbc.model.User;
 import jm.task.core.jdbc.util.Util;
+import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
-
+import java.util.Collections;
 import java.util.List;
 
 public class UserDaoHibernateImpl implements UserDao {
@@ -17,7 +18,7 @@ public class UserDaoHibernateImpl implements UserDao {
 
     @Override
     public void createUsersTable() {
-        Transaction transaction = null;
+        Transaction transaction;
         try (Session session = sessionFactory.getCurrentSession()) {
             transaction = session.beginTransaction();
             String sql = "CREATE TABLE IF NOT EXISTS users " +
@@ -26,26 +27,20 @@ public class UserDaoHibernateImpl implements UserDao {
                     "age TINYINT NOT NULL)";
             session.createSQLQuery(sql).executeUpdate();
             transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
+        } catch (HibernateException e) {
             e.printStackTrace();
         }
     }
 
     @Override
     public void dropUsersTable() {
-        Transaction transaction = null;
+        Transaction transaction;
         try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
             String sql = "DROP TABLE IF EXISTS users";
             session.createSQLQuery(sql).executeUpdate();
             transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
+        } catch (HibernateException e) {
             e.printStackTrace();
         }
     }
@@ -61,24 +56,26 @@ public class UserDaoHibernateImpl implements UserDao {
             transaction = session.beginTransaction();
             session.save(user);
             transaction.commit();
-        } catch (Exception e) {
+        } catch (HibernateException e) {
             if (transaction != null) {
                 transaction.rollback();
             }
             e.printStackTrace();
         }
     }
-
 
     @Override
     public void removeUserById(long id) {
         Transaction transaction = null;
         try (Session session = sessionFactory.openSession()) {
             User user = session.get(User.class, id);
+            if (user == null) {
+                return;
+            }
             transaction = session.beginTransaction();
             session.delete(user);
             transaction.commit();
-        } catch (Exception e) {
+        } catch (HibernateException e) {
             if (transaction != null) {
                 transaction.rollback();
             }
@@ -86,42 +83,26 @@ public class UserDaoHibernateImpl implements UserDao {
         }
     }
 
-//    @Override
-//    public List<User> getAllUsers() {
-//        List<User> userList = null;
-//        try (Session session = sessionFactory.openSession()) {
-//            String sql = "SELECT * FROM users";
-//            userList = session.createSQLQuery(sql).addEntity(User.class).list();
-//        } catch (Exception e) {
-//            e.printStackTrace();
-//        }
-//        return userList;
-//    }
-
     @Override
     public List<User> getAllUsers() {
-        List<User> userList = null;
+        List<User> userList = Collections.emptyList();
         try (Session session = sessionFactory.openSession()) {
             userList = session.createQuery("from User", User.class).getResultList();
-        } catch (Exception e) {
+        } catch (HibernateException e) {
             e.printStackTrace();
         }
         return userList;
     }
 
-
     @Override
     public void cleanUsersTable() {
-        Transaction transaction = null;
+        Transaction transaction;
         try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
             String sql = "TRUNCATE TABLE users";
             session.createSQLQuery(sql).executeUpdate();
             transaction.commit();
-        } catch (Exception e) {
-            if (transaction != null) {
-                transaction.rollback();
-            }
+        } catch (HibernateException e) {
             e.printStackTrace();
         }
     }
